Change production check to be portable

+++ osu.Server.Queues.ScoreStatisticsProcessor.Tests/DatabaseTest.cs
@@ -35,7 +35,8 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Tests
             {
                 // just a safety measure for now to ensure we don't hit production. since i was running on production until now.
                 // will throw if not on test database.
-                db.Query<int>("SELECT * FROM test_database");
+                if (db.QueryFirstOrDefault<int?>("SELECT * FROM osu_counts WHERE name = 'is_production'") != null)
+                    throw new InvalidOperationException("You are trying to do something very silly.");
 
                 db.Execute("TRUNCATE TABLE osu_user_stats");
                 db.Execute("TRUNCATE TABLE osu_user_stats_mania");
